Fix failure to set initially ungettable volume options like client.ssl
It's not clear why but attempting to get client.ssl before it is set
results in an error and an exit status of 2. This is different to
getting a fictional option like foo.bar, which results in a different
error and an exit status of 1.

Let's prevent an exit status of 2 from raising an exception and mark
the resource as non-existent in this case.

After resetting client.ssl, it returns to its ungettable state so do
not perform subsequent resets when the resource is non-existent.